refactor step 3: extract functions that do not refer to a single pr from pull_request_client
